+++ src/components/switch/switch.stories.tsx
@@ -58,7 +58,7 @@ export const Required: Story = () => {
   const [isChecked, setIsChecked] = useState(false);
   return (
     <Switch
-      label="Terms and Conditions"
+      label="Label"
       name="required"
       checked={isChecked}
       onChange={(e) => setIsChecked(e.target.checked)}
@@ -81,6 +81,20 @@ export const Reversed: Story = () => {
   );
 };
 Reversed.storyName = "Reversed";
+
+export const IsOptional: Story = () => {
+  const [isChecked, setIsChecked] = useState(false);
+  return (
+    <Switch
+      label="Label"
+      name="switch-name"
+      checked={isChecked}
+      isOptional
+      onChange={(e) => setIsChecked(e.target.checked)}
+    />
+  );
+};
+IsOptional.storyName = "IsOptional";
 
 export const Loading: Story = () => {
   return (