+++ src/locales/en-gb.ts
@@ -104,6 +104,9 @@ const enGB: Locale = {
   loader: {
     loading: () => "Loading",
   },
+  menuFullscreen: {
+    ariaLabels: { closeButton: () => "Close" },
+  },
   message: {
     closeButtonAriaLabel: () => "Close",
   },